Fix isTranslatable() for the user model
