+++ src/Models/Model.php
@@ -7,12 +7,13 @@ use A17\Twill\Models\Behaviors\HasPresenter;
 use Carbon\Carbon;
 use Cartalyst\Tags\TaggableInterface;
 use Cartalyst\Tags\TaggableTrait;
+use A17\Twill\Models\Behaviors\IsTranslatable;
 use Illuminate\Database\Eloquent\Model as BaseModel;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
 abstract class Model extends BaseModel implements TaggableInterface
 {
-    use HasPresenter, SoftDeletes, TaggableTrait;
+    use HasPresenter, SoftDeletes, TaggableTrait, IsTranslatable;
 
     public $timestamps = true;
 
@@ -60,30 +61,6 @@ abstract class Model extends BaseModel implements TaggableInterface
     public function scopeOnlyTrashed($query)
     {
         return $query->whereNotNull('deleted_at');
-    }
-
-    public function isTranslatable($columns = null)
-    {
-        // Model must have the trait
-        if (!classHasTrait($this, 'A17\Twill\Models\Behaviors\HasTranslation')) {
-            return false;
-        }
-
-        // Model must have the translatedAttributes property
-        if (!property_exists($this, 'translatedAttributes')) {
-            return false;
-        }
-
-        // If it's a check on certain columns
-        // They must be present in the translatedAttributes
-        if (filled($columns)) {
-            return collect($this->translatedAttributes)
-                ->intersect(collect($columns))
-                ->isNotEmpty();
-        }
-
-        // The translatedAttributes property must be filled
-        return collect($this->translatedAttributes)->isNotEmpty();
     }
 
     public function getFillable()
