+++ src/Models/User.php
@@ -5,6 +5,7 @@ namespace A17\Twill\Models;
 use A17\Twill\Models\Behaviors\HasMedias;
 use A17\Twill\Models\Behaviors\HasPresenter;
 use A17\Twill\Models\Enums\UserRole;
+use A17\Twill\Models\Behaviors\IsTranslatable;
 use A17\Twill\Notifications\Reset as ResetNotification;
 use A17\Twill\Notifications\Welcome as WelcomeNotification;
 use Illuminate\Auth\Authenticatable;
@@ -17,7 +18,7 @@ use PragmaRX\Google2FAQRCode\Google2FA;
 
 class User extends AuthenticatableContract
 {
-    use Authenticatable, Authorizable, HasMedias, Notifiable, HasPresenter, SoftDeletes;
+    use Authenticatable, Authorizable, HasMedias, Notifiable, HasPresenter, SoftDeletes, IsTranslatable;
 
     public $timestamps = true;
